[feat] : rvId타입 수정 #46

+++ demo/src/main/java/com/example/moviezip/dao/ReviewDao.java
@@ -15,7 +15,7 @@ public interface ReviewDao {
 
     List<Review> getReviewList(long mvId) throws DataAccessException;
 
-    Review getReviewDetail(long rvId) throws DataAccessException;
+    Review getReviewDetail(int rvId) throws DataAccessException;
 
     int updateReview(Review review) throws DataAccessException;
